feat: Default volumes and images when versions set
When users only set the `KernelVersion` and `OsVersion` fields, we
fill in the other fields required by flintlock based on those values.

For example, if given

```yaml
spec:
  template:
    spec:
      kernelVersion: 5.10.77
      osVersion: 1.23.5
```

We fill in the rest with:

```yaml
spec:
  template:
    spec:
      rootVolume:
        id: root
        image: ghcr.io/weaveworks-liquidmetal/capmvm-kubernetes:1.23.5
      kernel:
        filename: "boot/vmlinux"
        image: ghcr.io/weaveworks-liquidmetal/kernel-bin:5.10.77
      additionalVolumes:
      - id: modules
        image: ghcr.io/weaveworks-liquidmetal/kernel-modules:5.10.77
        mountPoint: /lib/modules/5.10.77
```

`convertToFlintlockAPI` has also been refactored to make it easier to
read, and tests have been added.